Swapped Mod for Better One
Changed out Phosphor, a mod which re-writes the lighting system, for Starlight. It promises better performance and is recommended by other mods and players. In my simple testing, it seemed to allow large light updates when changing the time of day to be instant, compared to a second of lag.